Remove sampling-based weighting in Client Side Stats (#8828)
## Summary of changes

Removes the incorrect sampling-based weighting from Client-Side stats

## Reason for change

When implementing the updated CSS implementation in #8420, we compared
with the canonicial Go implementation, and [added a
`GetWeight()`](https://github.com/DataDog/datadog-agent/blob/ce22e11ee71e55be717b9d9a3f8f3d7721a9c6d7/pkg/trace/stats/weight.go)
method based on this. However, the sampling rate that this is based on
is based on the deprecated `_sample_rate` tag (which .NET never
implemented), _not_ the normal sampling rate. Therefore this weighting
causes overcounting.

## Implementation details

- Rip out GetWeight() and associated usages
- Revert various doubles back to `long`s
- Remove the unnecessary stochastic rounding

## Test coverage

- Removed one test for `GetWeight()` behavior
- TBC a system test that would catch this - serverless e2e tests (thanks
@apiarian-datadog!) caught this one

+++ tracer/src/Datadog.Trace/Agent/StatsBucket.cs
@@ -22,21 +22,17 @@ namespace Datadog.Trace.Agent
 
         public StatsAggregationKey Key { get; }
 
-        // Based on https://github.com/DataDog/datadog-agent/blob/main/pkg/trace/stats/weight.go
-        // Hits, Errors, and TopLevelHits are doubles to accumulate fractional sampling weights (1/rate)
-        public double Hits { get; set; }
+        public long Hits { get; set; }
 
-        public double Errors { get; set; }
+        public long Errors { get; set; }
 
-        // Duration is a double to accumulate fractional sampling weights (1/rate), like Hits/Errors/TopLevelHits.
-        // Based on https://github.com/DataDog/datadog-agent/blob/main/pkg/trace/stats/statsraw.go
-        public double Duration { get; set; }
+        public long Duration { get; set; }
 
         public DDSketch OkSummary { get; }
 
         public DDSketch ErrorSummary { get; }
 
-        public double TopLevelHits { get; set; }
+        public long TopLevelHits { get; set; }
 
         public List<byte[]> PeerTags { get; }
